# services/inventory/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory API
  version: 0.7.4
  description: |
    Authenticated routes use Bearer JWT; `tenant_id` is taken from the token (ARCHITECTURE §6–§7), not `X-Tenant-Id`.
    JSON errors:  via pkg/common/errorcodes. JSON success: via pkg/common/httpresponse (`success`, `data`, `meta.request_id`, list `meta.pagination`).
    Probes `/health`, `/ready`, `/api/v1/inventory/health` use `text/plain`, not §9 JSON.
    POST `/api/v1/inventory/movements/{inbound|outbound|transfer|adjustment}` require header `Idempotency-Key` (ARCHITECTURE §9); replay with the same key and identical raw JSON body returns the existing draft movement.
paths:
...
  /api/v1/inventory/movements:
    get:
      operationId: getApiV1InventoryMovements
      tags: [movements]
      summary: List movements
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - name: type
          in: query
          schema:
            type: string
            enum: [inbound, outbound, transfer, adjustment]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, confirmed, cancelled]
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/inventory/movements/{movementId}:
    parameters:
      - $ref: "#/components/parameters/MovementId"
    get:
      operationId: getApiV1InventoryMovementsMovementId
      tags: [movements]
      summary: Get movement with lines
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Movement detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v1/inventory/movements/inbound:
    post:
      operationId: postApiV1InventoryMovementsInbound
      tags: [movements]
      summary: Create draft inbound movement
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundMovementCreateRequest"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/v1/inventory/movements/outbound:
    post:
      operationId: postApiV1InventoryMovementsOutbound
      tags: [movements]
      summary: Create draft outbound movement
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutboundMovementCreateRequest"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/v1/inventory/movements/transfer:
    post:
      operationId: postApiV1InventoryMovementsTransfer
      tags: [movements]
      summary: Create draft transfer movement
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferMovementCreateRequest"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/v1/inventory/movements/adjustment:
    post:
      operationId: postApiV1InventoryMovementsAdjustment
      tags: [movements]
      summary: Create draft adjustment movement
      description: Set exactly one of source_warehouse_id (decrease) or destination_warehouse_id (increase).
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustmentMovementCreateRequest"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/v1/inventory/movements/{movementId}/confirm:
    parameters:
      - $ref: "#/components/parameters/MovementId"
    post:
      operationId: postApiV1InventoryMovementsMovementIdConfirm
      tags: [movements]
      summary: Confirm draft movement (apply stock, audit, outbox)
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Confirmed movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/v1/inventory/movements/{movementId}/cancel:
    parameters:
      - $ref: "#/components/parameters/MovementId"
    post:
      operationId: postApiV1InventoryMovementsMovementIdCancel
      tags: [movements]
      summary: Cancel draft movement
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Cancelled movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementSuccessEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
...
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT; tenant scope from claims.
  parameters:
...
    MovementId:
      name: movementId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-supplied key for safe retries (ARCHITECTURE §9). Paired with a SHA-256 fingerprint of the raw JSON body on the server.
        Sending the same key with the same raw body returns the existing draft (no duplicate rows). Same key with a different body yields 409 `IDEMPOTENCY_CONFLICT`.
      schema:
        type: string
        maxLength: 255
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Search:
      name: search
      in: query
      schema:
        type: string
    Sort:
      name: sort
      in: query
      schema:
        type: string
        description: Field name to sort by (domain-specific)
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequestError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ConflictError:
      description: Resource conflict (e.g. duplicate name)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntityError:
      description: Domain rule violation (e.g. cannot delete category with active products)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Machine-readable error message
    PaginationMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
      properties:
        page:
          type: integer
          format: int32
        per_page:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    SuccessMeta:
      type: object
      description: §9 meta for JSON success responses (list endpoints include pagination).
      properties:
        request_id:
          type: string
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    MovementLine:
      type: object
      required:
        - id
        - movement_id
        - product_id
        - quantity
        - created_at
      properties:
        id:
          type: string
          format: uuid
        movement_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Movement:
      type: object
      required:
        - id
        - tenant_id
        - type
        - reference_number
        - created_by
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [inbound, outbound, transfer, adjustment]
        reference_number:
          type: string
        source_warehouse_id:
          type: string
          format: uuid
          nullable: true
        destination_warehouse_id:
          type: string
          format: uuid
          nullable: true
        created_by:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, confirmed, cancelled]
        notes:
          type: string
          nullable: true
        idempotency_key:
          type: string
          nullable: true
        idempotency_request_hash:
          type: string
          description: Lowercase SHA-256 hex of the raw JSON body used at create time; null for legacy rows.
          pattern: '^[0-9a-f]{64}$'
          nullable: true
        lines:
          type: array
          items:
            $ref: "#/components/schemas/MovementLine"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MovementLineCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
        notes:
          type: string
    MovementDraftFields:
      type: object
      required:
        - reference_number
        - lines
      properties:
        reference_number:
          type: string
        notes:
          type: string
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/MovementLineCreate"
    InboundMovementCreateRequest:
      allOf:
        - $ref: "#/components/schemas/MovementDraftFields"
        - type: object
          required:
            - destination_warehouse_id
          properties:
            destination_warehouse_id:
              type: string
              format: uuid
    OutboundMovementCreateRequest:
      allOf:
        - $ref: "#/components/schemas/MovementDraftFields"
        - type: object
          required:
            - source_warehouse_id
          properties:
            source_warehouse_id:
              type: string
              format: uuid
    TransferMovementCreateRequest:
      allOf:
        - $ref: "#/components/schemas/MovementDraftFields"
        - type: object
          required:
            - source_warehouse_id
            - destination_warehouse_id
          properties:
            source_warehouse_id:
              type: string
              format: uuid
            destination_warehouse_id:
              type: string
              format: uuid
    AdjustmentMovementCreateRequest:
      allOf:
        - $ref: "#/components/schemas/MovementDraftFields"
        - type: object
          properties:
            source_warehouse_id:
              type: string
              format: uuid
              nullable: true
            destination_warehouse_id:
              type: string
              format: uuid
              nullable: true
    MovementSuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/Movement"
        meta:
          $ref: "#/components/schemas/SuccessMeta"
    MovementListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
        meta:
          $ref: "#/components/schemas/SuccessMeta"
...